updated to requirements
The code takes two files from the user input: reference_beep and GoPro face camera. It calculates where the beep happens (in seconds) in the GoPro and crops out everything before the beep.
Then it spits out a cropped file of the GoPro camera with suffix "_shifted".
This cropped file should start at the same time when the 360 camera started.